A really dirty commit to be able to debug tests
Just to play it safe and irrevertible - a lot of logging functionality is added at the same time in the same commit.

Going back and forth for countless hours results in following: there is no way of telling why it did not work. It seems to be related to naming of test project, but I do not know what caused it. Since the number of references and variables between projects is vast in a C# project, there can be countless reasons for this. Now it works, but with a silly test project name. This will be fixed if I can find the error.