New regime for Xml Resources and fix of SendMessage-tests
The mock tests of the API uses FakeHttpHandlers which have xml 'prettified' within the code. Works once but is a pain to replace, like now when we are using v7 of API. Now, all responses are within separate files. Also, the fake-handlers should be created on the fly when needed, reducing the number of classes. This work starts now. SendMessage and Identify tests are fixed to use this pattern now.