Add fetch root entrypoint for REST hyperdriven api usage
Digipost has allways had a hyperdriven api. Thus hard coding the uri's
is a bad practice. Here I add the fetching of the root entrypoint and
use some relations from to fetch uri's for all implementent
functionality. (Inbox, Archive, Identify, search and message).

The entrypoint data is cached with a sliding cache eviction if not in use
for 5 min, but with a hard eviction on 1 hour. The reason for the hard
eviction is to be able to get a fresh public key for Digipost if Digipost
changes the public key.