Rework api to tak Typed Uri as parameters in sted of instances of objects with a given state.
Digipost Api is a Restful Api where the Uri are the actual resource. I makes
sense to use the resource as the parameter to fetch sertain data from the
api. In many cases the entrypoint has shortcuts to fetch more precisely the
data you want without first fetching other resources first and then find
resources from the response.
Fetching a specific ArchiveDocument by ID is such an example.
Others, like List InboxDocument are available from root. In thouse cases
you dont need to specifiy the resource uri since it is given from the root
element (Entrypoint). Fetching of the entry point is a transparent usecase
handled by the client library.

This is a breaking change of the api with regards to Inbox fetching.