Bypass issue with Newtonsoft.json
We use Newtonsoft to read the `secrets.json` file to get secrets
(sertifikat path and passord). Newtonsoft is an transitive dependency
through .NETStandard 2.0 (v 10.0.1). This is vulnerable if for DoS
due to improper handling of deeply nested json-files.
This should not usually be a problem since users of this client library
them selves craft the json file that is read. Anyway, we do the
mitigation recommended by by setting the MaxDepth to 128. This is quite
deep, and should not be an issue for anyone.
ref.: https://github.com/advisories/GHSA-5crp-9r3c-p9vr