Can send direct to print
The direct to print feature is a very specific case where we want the
message only to go to print without checking if a recipient has a
Digipost account or not.

To keep the api simple I created a new type PrintMessage that
excludes Digipost recipient addressing all together so that
the api should be easy to use.