Support Request for Registration
RfR er is method for delivery where you can wait some time for the
user to register and receive documents when they do. If they decide
to opt out there is an option for the sender to have the document printed
and sent by physical mail.

Generate code from new changes to xsd for request-for-registration