Add feature to get the document as stream or url to open in webbrowser
A Link relation changed it name in Digipost.

Changed name i api for type SharedDocumentRequestState -> ShareDocumentRequestState

Add support for Stop sharing

The ShareDocumentsRequestEvent data type can be added to the
ShareDocumentRequest. Fetch the ShareDocumentRequestState
and that will have a url to add Data to the request.

Fixes after rebase on main-branch

Use new Digipost DataType types for metadata and additional data.

This is a breaking change for how we use DataTypes in this library. Before
it was a stringy dependency where the user of the client library had to
Serialize the metadata them selves. Now this is done behind the scenes.
This will create better abstractions and type safety in addition to
better description of how metadata is supposed to be used.

* Separated smart post to a new header in the menu since the Send messages
  menu-point was getting very big.
* Added the documentation for the process for Share document request.

Use latest beta of data types

Clean up after rebase on main