Remove Sender as param to SetSenderInformation when org number and part id is used.
There was a unforseen confusion in the implementation where the Sender
actually was you as a broker (the holder of the authentication to Digipost api).
It used the Sender class as a parameter to use as the broker. This is
not needed since we can scope the request to the broker.

So GetSenderInformation can be fetched for the broker (specify no sender),
for a Sender you broker for (specify a sender id) and organisation
you broker for (org number and part id)