Use Digipost DataTypes v.4.0.0
This is a great improvement on handling smart post messages. Before
you had to create the metadata and serialize it your self. Now the
serialization is done internally in the client.

I have separated smart post into a separate documentation menu element
to have it more available and to shorten the menu point for Send messages.