Refactor IdentificationResult
The class IdentificationResult had a public constructor taking a string
as parameter and internally performed mapping between the generated
xml-domain and its own domain. This strange stringy contract made for
tight coupling between the domains. This should be done in the
transformation code DataTransferObjectConverter. Now with an internal
constructor and two different ones makes for better separation of consern
between what is an error and what is an identified person.

With regards to IdentificationError.Unknown vs NotFound, I made this comment in the code:

`Unknown` is never actually returned from Digipost. `NotFound` is. But since `Unknown` has
been the de facto response for `NotFound` since for ever in this client library
we should just keep it that way.