New document event: SHARE_DOCUMENTS_REQUEST_SHARING_WITHDRAWN
This event is generated whenever an active ShareDocumentsRequest is stopped
or the sharing is withdrawn. The same event is generated in either case.
This event is always generated for the Broker, not the Sender, so that
Brokers may poll events for a number of different Senders using the same
endpoint and parameters.